created context and introduction steps
