+++ App.tsx
@@ -2,13 +2,16 @@ import React from "react";
 
 import { SafeAreaProvider, SafeAreaView } from "react-native-safe-area-context";
 import AppWrapper from "./components/AppWrapper";
+import { UserProvider } from "./components/UserProvider";
 
 
 export default function App() {
   return (
     <SafeAreaProvider>
       <SafeAreaView style={{ flex: 1 }}>
-        <AppWrapper/>
+        <UserProvider>
+          <AppWrapper/>
+        </UserProvider>
       </SafeAreaView>
     </SafeAreaProvider>
   );
