+++ components/UserProvider/store.web.ts
@@ -1,5 +1,8 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
+
+// fetch / generate Key and pass it to function
+
 const storeKey = 'currentUser';
 
 export const getLocalStoreData = async () => {
